Move find_items to archive_lib
